requirements.txt
hidapi==0.14.0.post4
setuptools==75.6.0
